requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.2.6
opencv-python==4.12.0.88
pillow==12.0.0
Werkzeug==3.1.3
